backend/requirements.txt
fastapi
mangum
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
pydantic